CHROMIUM: soc: mediatek: SCPSYS: Drop parent/child relationship for mfg domains
The mfg power domains do not have a real parent/child relationship, so
the current implementation does not reflect reality. Nevertheless the
different mfg domains have hardware constraints in that they must be
enabled and disabled in a certain order.
To make that sure we no longer register the mfg_2d and mfg_async as
domains themselves but instead only the mfg domain which then also
enables/disabled the other two domains.

Note: This is a non-upstreamable CHROMIUM patch, because while it is
assumed that this is actually a work-around for an as-yet undiagnosed
power sequencing issue in the drm/img-rogue GPU driver.
An upstream solution would allow a fully functioning GPU driver to
also operate in the following "intermediate" power states:
 (a) mfg_async
 (b) mfg_async + mfg_2d

This fix in this patch disables these two modes since it always
enables/disables all three power domains together, mfg_async/mfg_2d/mfg.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>

BUG=chrome-os-partner:45713
TEST=build and boot oak-rev4
TEST=Let system idle over night
 => No messages like:
    "mtk-scpsys 10006000.scpsys: Failed to power off domain mfg"
TEST=Run power_LoadTest
 => No messages like:
    "mtk-scpsys 10006000.scpsys: Failed to power off domain mfg"

Change-Id: Id53ef23b0ec215a57c7ae1c9d17dc89b63baea33